requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
mysql-connector-python>=9.0.0
Werkzeug==2.3.7
Pillow==10.0.0
python-dotenv==1.0.0